unsigned long int ipv int gebruikt

+++ blok1/normaliser/normalise.cpp
@@ -27,22 +27,26 @@ int main(int argc, char** argv) {
         return -1;
     }
     int fs = in_info.samplerate;
-    int num_frames = in_info.frames;
-    int chn_num = in_info.channels;
+    unsigned long int num_frames = in_info.frames;
+    unsigned int chn_num = in_info.channels;
     printf("length of file: %fs\n", float(num_frames)/fs);
     printf("samplerate: %i\n", fs);
-    printf("num_frames: %i\n", num_frames);
+    printf("num_frames: %li\n", num_frames);
     printf("channels: %i\n", chn_num);
 
     float buffer[num_frames*chn_num];
-    int frames_read = sf_readf_float(in_file, buffer, num_frames);
-    printf("read %i frames form file\n", frames_read);
-    for(int i=0; i<frames_read; i++) {
-        printf("sample %i:\t%f\n", i, buffer[i]);
+    unsigned long int frames_read = sf_readf_float(in_file, buffer, num_frames);
+    printf("read %li frames form file\n", frames_read);
+
+    /*
+    for(unsigned long int i=0; i<frames_read; i++) {
+        printf("sample %li:\t%f\n", i, buffer[i]);
     }
+    */
+
     float max_val = 0; 
     float abs_val;
-    for(int n=0; n<num_frames*chn_num; n++) {
+    for(unsigned long int n=0; n<num_frames*chn_num; n++) {
         abs_val = std::sqrt(buffer[n]*buffer[n]);
         if(abs_val > max_val) {
             max_val = abs_val;
@@ -50,7 +54,7 @@ int main(int argc, char** argv) {
     }
     printf("peak value: %fdB\n", 20*std::log(max_val));
     printf("new max: %fdB\n", 20*std::log(max_val*gain));
-    for(int n=0; n<num_frames*chn_num; n++) {
+    for(unsigned long int n=0; n<num_frames*chn_num; n++) {
         buffer[n] = buffer[n]/max_val * gain;
     }
 
@@ -59,7 +63,7 @@ int main(int argc, char** argv) {
     out_name.append("_cp.wav");
     printf("string: %s\n", out_name.c_str());
     out_file = sf_open(out_name.c_str(), SFM_WRITE, &in_info);
-    int frames_wrtn = sf_writef_float(out_file, buffer, num_frames);
-    printf("wrote %i frames to file\n", frames_wrtn); 
+    unsigned long int frames_wrtn = sf_writef_float(out_file, buffer, num_frames);
+    printf("wrote %li frames to file\n", frames_wrtn); 
     return 0;
 }
